backend/requirements.txt
Django
djangorestframework 
pandas 
openpyxl 
django-cors-headers
xlrd
gunicorn
dj-database-url


asgiref
et_xmlfile
numpy
python-dateutil
python-dotenv
pytz
six
sqlparse
tzdata
whitenoise

djangorestframework-simplejwt
python-decouple
psycopg2-binary
redis
django-redis
django-extensions
Pillow
pyotp
qrcode
phonenumbers
wfastcgi

# Machine Learning Dependencies for Anomaly Detection
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
joblib>=1.3.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0

# AI Integration
google-generativeai>=0.3.0
